requirements.txt
appdirs==1.4.3
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
decorator==4.4.1
Flask==1.1.1
future==0.18.2
gmusicapi==12.1.1
gpsoauth==0.4.1
httplib2==0.15.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.2
MarkupSafe==1.1.1
MechanicalSoup==0.12.0
mock==3.0.5
mutagen==1.43.0
oauth2client==4.1.3
proboscis==1.2.6.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pycryptodomex==3.9.4
python-dateutil==2.8.1
python-telegram-bot==12.2.0
requests==2.22.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
tornado==6.0.3
urllib3==1.25.7
validictory==1.1.2
Werkzeug==0.16.0
